backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
qdrant-client
google-generativeai
httpx
aiofiles
pillow
python-dotenv
numpy
sqlalchemy
aiosqlite
